Refactor tests + exclusion des tests d'intégration de la CI
