+++ .github/workflows/ci.yml
@@ -35,7 +35,7 @@ jobs:
 
     - name: Run tests with coverage
       run: |
-        pytest --cov=app --cov-report=xml
+        pytest --cov=app --cov-report=xml -m "not integration"
       env:
         PYTHONPATH: .
